+++ cmn-data-svc-server/src/test/java/org/acumos/cds/test/CdsRepositoryServiceTest.java
@@ -30,6 +30,7 @@ import javax.validation.ConstraintViolationException;
 
 import org.acumos.cds.CodeNameType;
 import org.acumos.cds.domain.MLPArtifact;
+import org.acumos.cds.domain.MLPCatRoleMap;
 import org.acumos.cds.domain.MLPCatSolMap;
 import org.acumos.cds.domain.MLPCatalog;
 import org.acumos.cds.domain.MLPCodeNamePair;
@@ -74,6 +75,7 @@ import org.acumos.cds.domain.MLPUserRoleMap;
 import org.acumos.cds.repository.ArtifactRepository;
 import org.acumos.cds.repository.CatSolMapRepository;
 import org.acumos.cds.repository.CatalogRepository;
+import org.acumos.cds.repository.CatRoleMapRepository;
 import org.acumos.cds.repository.CommentRepository;
 import org.acumos.cds.repository.CompSolMapRepository;
 import org.acumos.cds.repository.DocumentRepository;
@@ -148,6 +150,8 @@ public class CdsRepositoryServiceTest {
 	private ArtifactRepository artifactRepository;
 	@Autowired
 	private CatalogRepository catalogRepository;
+	@Autowired
+	private CatRoleMapRepository catalogRoleMapRepository;
 	@Autowired
 	private CatalogSearchService catalogSearchService;
 	@Autowired
@@ -400,9 +404,12 @@ public class CdsRepositoryServiceTest {
 			logger.info("User Info {}", cu);
 			userRoleMapRepository.save(new MLPUserRoleMap(cu.getUserId(), cr2.getRoleId()));
 
-			long usersInRole = userRoleMapRepository.countRoleUsers(cr2.getRoleId());
-			Assert.assertNotEquals(0, usersInRole);
-			logger.info("Count of users in role: {}", usersInRole);
+			long usersInRoleCount = userRoleMapRepository.countRoleUsers(cr2.getRoleId());
+			Assert.assertNotEquals(0, usersInRoleCount);
+			logger.info("Count of users in role: {}", usersInRoleCount);
+
+			Page<MLPUser> usersInRole = userRoleMapRepository.findUsersByRoleId(cr2.getRoleId(), PageRequest.of(0, 5));
+			Assert.assertNotEquals(0, usersInRole.getNumberOfElements());
 
 			logger.info("Checking role content");
 			Optional<MLPRole> opt = roleRepository.findById(cr2.getRoleId());
@@ -1137,6 +1144,11 @@ public class CdsRepositoryServiceTest {
 	@Test
 	public void testRoleAndFunctions() throws Exception {
 		try {
+			MLPCatalog cat = new MLPCatalog("RS", true, "name", "pname", "http://pub.org");
+			cat = catalogRepository.save(cat);
+			Assert.assertNotNull("Catalog ID", cat.getCatalogId());
+			logger.info("Created restricted catalog {}", cat);
+
 			MLPUser cu = null;
 			cu = new MLPUser();
 			cu.setActive(true);
@@ -1160,12 +1172,6 @@ public class CdsRepositoryServiceTest {
 			Page<MLPRole> emptyRoles = roleSearchService.findRoles("bogus", Boolean.TRUE, false, PageRequest.of(0, 5));
 			Assert.assertTrue(emptyRoles.isEmpty());
 
-			logger.info("Assigning role to user");
-			userRoleMapRepository.save(new MLPUserRoleMap(cu.getUserId(), cr.getRoleId()));
-
-			Iterable<MLPRole> roles = roleRepository.findByUser(cu.getUserId());
-			Assert.assertTrue(roles.iterator().hasNext());
-
 			MLPRoleFunction crf = new MLPRoleFunction();
 			final String roleFuncName = "My test role function";
 			crf.setName(roleFuncName);
@@ -1182,11 +1188,35 @@ public class CdsRepositoryServiceTest {
 			MLPRoleFunction roleFuncOne = resrf.iterator().next();
 			Assert.assertEquals(roleFuncName, roleFuncOne.getName());
 
+			logger.info("Assigning role to catalog");
+			catalogRoleMapRepository.save(new MLPCatRoleMap(cat.getCatalogId(), cr.getRoleId()));
+			Iterable<MLPRole> catRoles = roleRepository.findByCatalog(cat.getCatalogId());
+			Assert.assertTrue(catRoles.iterator().hasNext());
+
+			logger.info("Assigning role to user");
+			userRoleMapRepository.save(new MLPUserRoleMap(cu.getUserId(), cr.getRoleId()));
+			Iterable<MLPRole> userRoles = roleRepository.findByUser(cu.getUserId());
+			Assert.assertTrue(userRoles.iterator().hasNext());
+
+			long catalogsInRoleCount = catalogRoleMapRepository.countRoleCatalogs(cr.getRoleId());
+			Assert.assertNotEquals(0, catalogsInRoleCount);
+			logger.info("Count of catalogs in role: {}", catalogsInRoleCount);
+
+			Page<MLPCatalog> catalogsInRole = catalogRoleMapRepository.findCatalogsByRoleId(cr.getRoleId(),
+					PageRequest.of(0, 5));
+			Assert.assertNotEquals(0, catalogsInRole.getNumberOfElements());
+
+			logger.info("Removing catalog from role");
+			catalogRoleMapRepository.deleteById(new MLPCatRoleMap.CatalogRoleMapPK(cat.getCatalogId(), cr.getRoleId()));
+
 			logger.info("Removing user from role");
 			userRoleMapRepository.deleteById(new MLPUserRoleMap.UserRoleMapPK(cu.getUserId(), cr.getRoleId()));
 
 			logger.info("Deleting test user");
 			userRepository.deleteById(cu.getUserId());
+
+			logger.info("Deleting test catalog");
+			catalogRepository.deleteById(cat.getCatalogId());
 
 			logger.info("Deleting test role function");
 			roleFunctionRepository.delete(roleFuncOne);