+++ docs/release-notes.rst
@@ -28,6 +28,7 @@ Version 3.1.0, 3 Dec 2019
 -------------------------
 * Drop Right To Use entities and controller (`ACUMOS-3460 <https://jira.acumos.org/browse/ACUMOS-3460>`_)
 * Add C++ and Spark toolkit types (`ACUMOS-3726 <https://jira.acumos.org/browse/ACUMOS-3726>`_)
+* Add catalog-role mappings (`ACUMOS-3681 <https://jira.acumos.org/browse/ACUMOS-3681>`_)
 * Upgrade to Spring-Boot version 2.1.10.RELEASE
 * Requires database schema version 3.1